initialization again

+++ libsponge/byte_stream.hh
@@ -18,10 +18,10 @@ class ByteStream {
     // that's a sign that you probably want to keep exploring
     // different approaches.
     size_t stream_capacity;
-    size_t total_written;
-    size_t total_read;
-    size_t curr_size; //size of written buffer
-    size_t curr_left_size; // remaining size
+    int total_written;
+    int total_read;
+    int curr_size; //size of written buffer
+    int curr_left_size; // remaining size
     bool the_end;
     
     ostringstream buffer;
